backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2
